+++ CHANGELOG.md
@@ -21,8 +21,9 @@ and this project adheres to
 - 🚚(frontend) rename "wellknown" directory to "well-known" #1009
 - 🌐(frontend) localize SR modifier labels #1010
 - ⬆️(backend) update python dependencies #1011
-- ♿️(a11y) fix focus ring on tab container components
+- ♿️(a11y) fix focus ring on tab container components #1012
 
+- ♻️(frontend) use react aria toolbar for arrow key navigation #1005
 
 ## [1.8.0] - 2026-02-20